requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile - -o requirements.txt
aiohappyeyeballs==2.4.8
    # via aiohttp
aiohttp==3.11.13
    # via
    #   datasets
    #   fsspec
aiosignal==1.3.2
    # via aiohttp
aiosqlite==0.21.0
    # via tortoise-orm
annotated-types==0.7.0
    # via pydantic
anyio==4.8.0
    # via
    #   httpx
    #   openai
attrs==25.1.0
    # via aiohttp
certifi==2025.1.31
    # via
    #   httpcore
    #   httpx
    #   requests
    #   sentry-sdk
charset-normalizer==3.4.1
    # via requests
ciso8601==2.3.2
click==8.1.8
contourpy==1.3.1
    # via matplotlib
cycler==0.12.1
    # via matplotlib
datasets==3.3.2
dill==0.3.8
    # via
    #   datasets
    #   multiprocess
distro==1.9.0
    # via openai
docker-pycreds==0.4.0
docstring-parser==0.16
    # via typed-argument-parser
dotenv==0.9.9
filelock==3.17.0
    # via
    #   datasets
    #   huggingface-hub
    #   transformers
fonttools==4.56.0
    # via matplotlib
frozenlist==1.5.0
    # via
    #   aiohttp
    #   aiosignal
fsspec==2024.12.0
    # via
    #   datasets
    #   huggingface-hub
gitdb==4.0.12
    # via gitpython
gitpython==3.1.44
h11==0.14.0
    # via httpcore
httpcore==1.0.7
    # via httpx
httpx==0.28.1
    # via openai
huggingface-hub==0.29.2
    # via
    #   datasets
    #   tokenizers
    #   transformers
idna==3.10
    # via
    #   anyio
    #   httpx
    #   requests
    #   yarl
iso8601==2.1.0
    # via tortoise-orm
jiter==0.8.2
    # via openai
joblib==1.4.2
    # via scikit-learn
kiwisolver==1.4.8
    # via matplotlib
loguru==0.7.3
matplotlib==3.10.1
    # via seaborn
multidict==6.1.0
    # via
    #   aiohttp
    #   yarl
multiprocess==0.70.16
    # via datasets
mypy-extensions==1.0.0
    # via typing-inspect
numpy==2.2.3
    # via
    #   contourpy
    #   datasets
    #   matplotlib
    #   pandas
    #   scikit-learn
    #   scipy
    #   seaborn
    #   transformers
openai==1.65.3
orjson==3.10.15
packaging==24.2
    # via
    #   datasets
    #   huggingface-hub
    #   matplotlib
    #   transformers
    #   typed-argument-parser
pandas==2.2.3
    # via
    #   datasets
    #   seaborn
pillow==11.1.0
    # via matplotlib
platformdirs==4.3.6
propcache==0.3.0
    # via
    #   aiohttp
    #   yarl
protobuf==5.29.3
psutil==7.0.0
pyarrow==19.0.1
    # via datasets
pydantic==2.10.6
    # via openai
pydantic-core==2.27.2
    # via pydantic
pyparsing==3.2.1
    # via matplotlib
pypika-tortoise==0.5.0
    # via tortoise-orm
python-dateutil==2.9.0.post0
    # via
    #   matplotlib
    #   pandas
python-dotenv==1.0.1
    # via dotenv
pytz==2025.1
    # via
    #   pandas
    #   tortoise-orm
pyyaml==6.0.2
    # via
    #   datasets
    #   huggingface-hub
    #   transformers
regex==2024.11.6
    # via transformers
requests==2.32.3
    # via
    #   datasets
    #   huggingface-hub
    #   transformers
result==0.17.0
ruff==0.9.9
safetensors==0.5.3
    # via transformers
scikit-learn==1.6.1
scipy==1.15.2
    # via scikit-learn
seaborn==0.13.2
sentry-sdk==2.22.0
setproctitle==1.3.5
setuptools==76.0.0
six==1.17.0
    # via
    #   docker-pycreds
    #   python-dateutil
smmap==5.0.2
    # via gitdb
sniffio==1.3.1
    # via
    #   anyio
    #   openai
threadpoolctl==3.6.0
    # via scikit-learn
tokenizers==0.21.0
    # via transformers
tortoise-orm==0.24.2
tqdm==4.67.1
    # via
    #   datasets
    #   huggingface-hub
    #   openai
    #   transformers
transformers==4.49.0
typed-argument-parser==1.10.1
typing-extensions==4.12.2
    # via
    #   aiosqlite
    #   anyio
    #   huggingface-hub
    #   openai
    #   pydantic
    #   pydantic-core
    #   typing-inspect
typing-inspect==0.9.0
    # via typed-argument-parser
tzdata==2025.1
    # via pandas
urllib3==2.3.0
    # via
    #   requests
    #   sentry-sdk
uvloop==0.21.0
xxhash==3.5.0
    # via datasets
yarl==1.18.3
    # via aiohttp
